requirements.txt
numpy==1.26.4
scipy==1.11.4
scikit-learn==1.3.2
sentence-transformers==2.7.0

langchain==0.2.14
langchain-community==0.2.12
langchain-openai==0.1.22
langchain-huggingface

chromadb==0.5.5
python-dotenv==1.0.1
requests

fastapi
uvicorn
sqlalchemy
pymysql